pyproject.toml
[project]
name = "workmate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.5.0",
    "google-generativeai>=0.8.6",
    "google-genai>=1.64.0",
    "langchain-text-splitters>=1.1.0",
    "python-dotenv>=1.2.1",
    "fastapi>=0.131.0",
    "httpx>=0.28.1",
    "pydantic-settings>=2.13.1",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0.46",
    "uvicorn[standard]>=0.40.0",
    "python-multipart>=0.0.22",
    "pypdf2>=3.0.1",
    "sse-starlette>=3.3.2",
    "voyageai>=0.3.7",
    "bm25s>=0.2.12",
    "psycopg2-binary>=2.9.10",
]
